Rename massage form file and update imports
