+++ portfolio/src/components/home/module/first_reply.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react"
 
-import MessageFormProps from "./massage_form"
+import MessageFormProps from "./message_form"
 
 import Reply from "./return_respond"
 import { default_second_message_ja, default_second_message_en } from "./data"
