+++ portfolio/src/components/home/module/return_respond.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import MessageFormProps from './massage_form';
+import MessageFormProps from './message_form';
 
 import { default_second_message_ja, default_second_message_en } from './data';
